Deleted CommonUtil.java and restored JavaUtil to previous state.
